Centralized ZenML repository
* `zenml init` is deprecated. The global Repository is now initialized
automatically upon instantiation in the ~/.config/zenml location
* the Repository is made into a singleton, to avoid multiple instances
* the GlobalConfig is made into a singleton, to avoid multiple instances
* moved `version` from repository to global configuration, because it is
an attribute that needs to be addressed globally e.g. when migrating the
global configuration to a new version. Implemented rudimentary schema migration
and validation for the global configuration.
* implemented configuration profiles and replaced RepositoryConfiguration with
ConfigProfile
* replaced support for .zen folders as reference points for building
container images and loading source path classes with dynamic detection
of root python paths